*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/irda/code/trunk@15 a0daaf5f-cf0d-0410-a784-93f8423a4122

+++ irda-utils/irdadump/src/netbuf.h
@@ -65,7 +65,7 @@ void     g_netbuf_set_size(GNetBuf *msg, guint len);
 void     g_netbuf_print(GNetBuf *msg);
 
 static inline guint8 *g_netbuf_get_data(GNetBuf *msg) { return msg->data; }
-static inline guint8 g_netbuf_get_len(GNetBuf *msg) { return msg->len; }
+static inline guint g_netbuf_get_len(GNetBuf *msg) { return msg->len; }
 
 #endif
